moved enterprise compaction test
