+++ producer/src/test/kotlin/streams/integrations/KafkaEventRouterSuiteIT.kt
@@ -17,8 +17,7 @@ import streams.utils.StreamsUtils
         KafkaEventRouterSimpleTSE::class,
         KafkaEventRouterWithConstraintsTSE::class,
         KafkaEventRouterEnterpriseTSE::class,
-        KafkaEventRouterLogCompactionTSE::class,
-        KafkaEventRouterEnterpriseLogCompactionTSE::class
+        KafkaEventRouterLogCompactionTSE::class
 )
 class KafkaEventRouterSuiteIT {
